Keep the overflow record a column read alone could not see
A common compound's overflow record can stand under a tile column that is no
longer saturated: extraction, growth and diffusion write the column directly,
and `tileMatterAmount` reads the column alone while it is under the cap. A
caller that set what it read plus a little past the cap had never seen the
record, and `setTileMatterAmount` wrote the overflow over it — measured on
battery causal-origin at tick 18847, a nutrient tile at 65,532 with 641 on
record set to 65,536 and left with 1, six hundred and forty gone. What the
caller could not have read now stays on record. The camp-abandon deposit
clamped a record at 65,535 with `u16`; a record is a number, and it adds.

The cohort smoke sets a saturated tile a little past the cap from a column
read alone and asserts the record and the total both survive.

+++ src/game/sections/25-settlements-construction.js
@@ -320,7 +320,7 @@ function updateSettlements() {
       for (let sp = 0; sp < SPECIES_COUNT; sp++) {
         const amount = c.inventory[sp] + c.structure.composition[sp];
         if (sp < COMMON_CHEM) W.tiles.chem[sp][ti] = u16(W.tiles.chem[sp][ti] + amount);
-        else W.tiles.rareChem[`${ti}:${sp}`] = u16((W.tiles.rareChem[`${ti}:${sp}`] || 0) + amount);
+        else W.tiles.rareChem[`${ti}:${sp}`] = (W.tiles.rareChem[`${ti}:${sp}`] || 0) + amount;
       }
       emitEvent("CampAbandonedEvent", {
         subjects: [c.entityId],
